[Cleanup] Remove deprecated OpenSSL Bignum support
We've been using GMP for Bignum support by default already, and now that
 zerocoin has been replaced, there is little reason to keep the OpenSSL
 fallback Bignum support anymore.
