+++ build-aux/m4/gmp.m4
@@ -9,7 +9,6 @@ if test x"$has_gmp" != x"yes"; then
     AC_CHECK_LIB(gmp, __gmpz_init,[
       has_gmp=yes;
       LIBS="$LIBS -lgmp";
-      AC_DEFINE(HAVE_LIBGMP, 1, [Define this symbol if libgmp is installed])
     ])
   ])
 fi
