+++ src/libzerocoin/bignum.h
@@ -11,12 +11,7 @@
 #include "config/pivx-config.h"
 #endif
 
-#if defined(USE_NUM_OPENSSL)
-#include <openssl/bn.h>
-#endif
-#if defined(USE_NUM_GMP)
 #include <gmp.h>
-#endif
 
 #include <stdexcept>
 #include <vector>
@@ -37,12 +32,7 @@ public:
 /** C++ wrapper for BIGNUM */
 class CBigNum
 {
-#if defined(USE_NUM_OPENSSL)
-    BIGNUM* bn;
-#endif
-#if defined(USE_NUM_GMP)
     mpz_t bn;
-#endif
 public:
     CBigNum();
     CBigNum(const CBigNum& b);
@@ -169,12 +159,7 @@ public:
     *                          default causes error rate of 2^-80.
     * @return true if prime
     */
-#if defined(USE_NUM_OPENSSL)
-    bool isPrime(const int checks=BN_prime_checks) const;
-#endif
-#if defined(USE_NUM_GMP)
     bool isPrime(const int checks=15) const;
-#endif
 
     bool isOne() const;
     bool operator!() const;
@@ -205,91 +190,6 @@ public:
     friend inline bool operator>(const CBigNum& a, const CBigNum& b);
 };
 
-#if defined(USE_NUM_OPENSSL)
-class CAutoBN_CTX
-{
-protected:
-    BN_CTX* pctx;
-    BN_CTX* operator=(BN_CTX* pnew) { return pctx = pnew; }
-
-public:
-    CAutoBN_CTX()
-    {
-        pctx = BN_CTX_new();
-        if (pctx == NULL)
-            throw bignum_error("CAutoBN_CTX : BN_CTX_new() returned NULL");
-    }
-
-    ~CAutoBN_CTX()
-    {
-        if (pctx != NULL)
-            BN_CTX_free(pctx);
-    }
-
-    operator BN_CTX*() { return pctx; }
-    BN_CTX& operator*() { return *pctx; }
-    BN_CTX** operator&() { return &pctx; }
-    bool operator!() { return (pctx == NULL); }
-};
-
-inline const CBigNum operator+(const CBigNum& a, const CBigNum& b) {
-    CBigNum r;
-    if (!BN_add(r.bn, a.bn, b.bn))
-        throw bignum_error("CBigNum::operator+ : BN_add failed");
-    return r;
-}
-inline const CBigNum operator-(const CBigNum& a, const CBigNum& b) {
-    CBigNum r;
-    if (!BN_sub(r.bn, a.bn, b.bn))
-        throw bignum_error("CBigNum::operator- : BN_sub failed");
-    return r;
-}
-inline const CBigNum operator-(const CBigNum& a) {
-    CBigNum r(a);
-    BN_set_negative(r.bn, !BN_is_negative(r.bn));
-    return r;
-}
-inline const CBigNum operator*(const CBigNum& a, const CBigNum& b) {
-    CAutoBN_CTX pctx;
-    CBigNum r;
-    if (!BN_mul(r.bn, a.bn, b.bn, pctx))
-        throw bignum_error("CBigNum::operator* : BN_mul failed");
-    return r;
-}
-inline const CBigNum operator/(const CBigNum& a, const CBigNum& b) {
-    CAutoBN_CTX pctx;
-    CBigNum r;
-    if (!BN_div(r.bn, NULL, a.bn, b.bn, pctx))
-        throw bignum_error("CBigNum::operator/ : BN_div failed");
-    return r;
-}
-inline const CBigNum operator%(const CBigNum& a, const CBigNum& b) {
-    CAutoBN_CTX pctx;
-    CBigNum r;
-    if (!BN_nnmod(r.bn, a.bn, b.bn, pctx))
-        throw bignum_error("CBigNum::operator% : BN_div failed");
-    return r;
-}
-inline const CBigNum operator<<(const CBigNum& a, unsigned int shift) {
-    CBigNum r;
-    if (!BN_lshift(r.bn, a.bn, shift))
-        throw bignum_error("CBigNum:operator<< : BN_lshift failed");
-    return r;
-}
-inline const CBigNum operator>>(const CBigNum& a, unsigned int shift) {
-    CBigNum r = a;
-    r >>= shift;
-    return r;
-}
-inline bool operator==(const CBigNum& a, const CBigNum& b) { return (BN_cmp(a.bn, b.bn) == 0); }
-inline bool operator!=(const CBigNum& a, const CBigNum& b) { return (BN_cmp(a.bn, b.bn) != 0); }
-inline bool operator<=(const CBigNum& a, const CBigNum& b) { return (BN_cmp(a.bn, b.bn) <= 0); }
-inline bool operator>=(const CBigNum& a, const CBigNum& b) { return (BN_cmp(a.bn, b.bn) >= 0); }
-inline bool operator<(const CBigNum& a, const CBigNum& b)  { return (BN_cmp(a.bn, b.bn) < 0); }
-inline bool operator>(const CBigNum& a, const CBigNum& b)  { return (BN_cmp(a.bn, b.bn) > 0); }
-#endif
-
-#if defined(USE_NUM_GMP)
 inline const CBigNum operator+(const CBigNum& a, const CBigNum& b) {
     CBigNum r;
     mpz_add(r.bn, a.bn, b.bn);
@@ -336,7 +236,6 @@ inline bool operator<=(const CBigNum& a, const CBigNum& b) { return (mpz_cmp(a.b
 inline bool operator>=(const CBigNum& a, const CBigNum& b) { return (mpz_cmp(a.bn, b.bn) >= 0); }
 inline bool operator<(const CBigNum& a, const CBigNum& b)  { return (mpz_cmp(a.bn, b.bn) < 0); }
 inline bool operator>(const CBigNum& a, const CBigNum& b)  { return (mpz_cmp(a.bn, b.bn) > 0); }
-#endif
 
 inline std::ostream& operator<<(std::ostream &strm, const CBigNum &b) { return strm << b.ToString(10); }
 
@@ -348,4 +247,4 @@ const CBigNum BN_ONE = CBigNum(1);
 const CBigNum BN_TWO = CBigNum(2);
 const CBigNum BN_THREE = CBigNum(3);
 
-#endif
+#endif // BITCOIN_BIGNUM_H
